pyproject.toml
[tool.poetry]
name = "label-craft"
version = "0.1.0"
description = "Label craft"
authors = ["Alexander Sokolov <sokolov23al@gmail.com>"]
packages = [{include = "label_craft"}]

[tool.poetry.dependencies]
python = "^3.9"
dvc = "^3.0.0"
hydra-core = "^1.3.0"
requests = "^2.31.0"
omegaconf = "^2.3.0"
loguru = "^0.7.3"
pytorch-lightning = "^2.0.0"
transformers = "^4.30.0"
torch = "^2.0.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
pyarrow = ">=4.0.0,<20"
numpy = "<2.0.0"
tensorboard = "^2.15.0"
mlflow = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pre-commit = "^3.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
train = "label_craft.train:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["label_craft"]
known_third_party = ["torch", "pytorch_lightning", "transformers", "pandas", "numpy", "sklearn", "loguru", "mlflow", "hydra", "omegaconf"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg",
    ".mypy_cache",
    ".pytest_cache",
    ".coverage",
    "htmlcov",
    "mlruns",
    "logs"
]
per-file-ignores = [
    "__init__.py:F401"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"